feat: implement modules mjs
